Перенес функции, отвечающие за редактирование карточки ( добавление/изменение картинки, добавленик/удаление пользовательских примеров )
