+++ engLearn/views.py
@@ -61,32 +61,5 @@ class WordsDetailView(DetailView):
         return queryset.order_by('id')
 
 
-def add_examples(request, word_slug):
-    word = Words.objects.get(slug=word_slug)
-    if request.method == 'POST':
-        form = WordExamplesForm(request.POST)
-        if form.is_valid():
-            en_example_user = form.cleaned_data["en_example_user"]
-            ru_example_user = form.cleaned_data["ru_example_user"]
-            WordExamples.objects.create(user=request.user, word=word,
-                                        en_example_user=en_example_user,
-                                        ru_example_user=ru_example_user)
-            return redirect('word_detail', word_slug=word.slug)
-    else:
-        form = WordExamplesForm()
-    return render(request, 'engLearn/add_examples.html', {'word': word, 'form': form})
 
 
-def change_image(request, word_slug):
-    word = Words.objects.get(slug=word_slug)
-    if request.method == 'POST':
-        form = ChangeImageForm(request.POST, request.FILES)
-        if form.is_valid():
-            image = form.cleaned_data['image']
-            obj, created = WordImageUser.objects.get_or_create(word=word, user=request.user)
-            obj.image = image
-            obj.save()
-            return redirect('word_detail', word_slug=word.slug)
-    else:
-        form = ChangeImageForm()
-    return render(request, 'engLearn/change_image.html', {'word': word, 'form': form})
